requirements.txt

channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
django-import-export==3.3.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
MarkupPy==1.14
openpyxl==3.1.2
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
environs==10.0.0
pipwin==0.5.2
